Fixed connection issue
Problem was found to be an issue with the directory structure.

+++ back-end/src/main/java/com/crypto_tutor/repositories/UserRepository.java
@@ -1,9 +1,9 @@
 package com.crypto_tutor.repositories;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.crypto_tutor.models.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
-    
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
 }
